JwtTokenServiceTests

+++ LeonCam2/Models/Settings.cs
@@ -6,12 +6,12 @@ namespace LeonCam2.Models
     {
         public string JwtKey { get; set; }
 
-        public int JwtTokenLifeTimeInHours { get; set; } = 24;
+        public double JwtTokenLifeTimeInHours { get; set; } = 24;
 
-        public int BlockTimeInMinutes { get; set; } = 15;
+        public double BlockTimeInMinutes { get; set; } = 15;
 
         public int MaxNumberOfLoginAttempts { get; set; } = 3;
 
-        public int BlackListControlIntervalInHours { get; set; } = 24;
+        public double BlackListControlIntervalInHours { get; set; } = 24;
     }
 }
